frontend/requirements.txt
gradio>=4.0.0
requests==2.32.5
pillow==12.1.0
pandas==2.3.3
